:construction: section date settings modal
Refactored some of the initialization settings in HtmlEditor so that the modifications to the editor happen in the plugin files, rather than in the components. Basically the "open" callbacks now return a promise so buttons that trigger them in /plugin can then be responsible for making the content modifications.

Instead of passing the entire editor to the modal callbacks, pass an object of data that you would like to be present on the modal.

editor.settings.ordercloud.open_section_date_settings is a new modal for setting [data-oc-start-date] and [data-oc-end-date] attributes on section template containers. I'm not 100% sure where the script to hide/show based on these attributes will live but my guess would be in the Page Renderer component.